Refactore Package Modules
- Remove brdige table package_feature_setting to minimize joins for performance sake.
