+++ modules/Subscription/Models/Feature.php
@@ -7,6 +7,7 @@ namespace Modules\Subscription\Models;
 use Illuminate\Database\Eloquent\Model;
 use BasePackage\Shared\Traits\UuidTrait;
 use BasePackage\Shared\Traits\BaseFilterable;
+use Modules\Subscription\Enums\FeatureLimitTypeEnum;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Feature extends Model
@@ -22,12 +23,19 @@ class Feature extends Model
     protected $fillable = [
         'name',
         'slug',
-        'module_id'
+        'module_id',
+        'package_id',
+        'is_enabled',
+        'limit',
+        'limit_type',
     ];
 
     protected $casts = [
         'id' => 'string',
         'name' => 'json',
+        'is_enabled' => 'boolean',
+        'limit' => 'integer',
+        'limit_type' => FeatureLimitTypeEnum::class,
     ];
 
     /**
@@ -38,5 +46,15 @@ class Feature extends Model
     public function module(): BelongsTo
     {
         return $this->belongsTo(Module::class);
+    }
+
+     /**
+     * Get the package that owns this setting.
+     *
+     * @return BelongsTo<Package>
+     */
+    public function package(): BelongsTo
+    {
+        return $this->belongsTo(Package::class);
     }
 }
