--- modules/Subscription/Models/Subscription.php
@@ -1,40 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Modules\Subscription\Models;
-
-use BasePackage\Shared\Traits\UuidTrait;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Modules\Subscription\Database\factories\SubscriptionFactory;
-use BasePackage\Shared\Traits\BaseFilterable;
-//use BasePackage\Shared\Traits\HasTranslations;
-
-class Subscription extends Model
-{
-    use HasFactory;
-    use UuidTrait;
-    use BaseFilterable;
-    //use HasTranslations;
-    //use SoftDeletes;
-
-    //public array $translatable = [];
-
-    public $incrementing = false;
-
-    protected $keyType = 'string';
-
-    protected $fillable = [
-        'name',
-    ];
-
-    protected $casts = [
-        'id' => 'string',
-    ];
-
-    protected static function newFactory(): SubscriptionFactory
-    {
-        return SubscriptionFactory::new();
-    }
-}
